# civicswipe/docs/api-spec.yaml
openapi: 3.1.0
info:
  title: CivicSwipe API
  version: 1.0.0
  description: |
    API for CivicSwipe - A legislation tracking and voting app.
    Phoenix, Arizona MVP Release.
    
    ## Authentication
    All endpoints except /auth/* require Bearer JWT token:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Pagination
    Cursor-based pagination using `cursor` and `limit` query parameters.
    
    ## Idempotency
    POST endpoints that record votes support `Idempotency-Key` header.
servers:
  - url: https://api.civicswipe.app/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Local development
...
paths:
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Create new user account (requires address)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, address]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: StrongPassword123!
                address:
                  $ref: '#/components/schemas/Address'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    $ref: '#/components/schemas/Tokens'
                  location:
                    $ref: '#/components/schemas/LocationResolution'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens:
                    $ref: '#/components/schemas/Tokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens:
                    $ref: '#/components/schemas/Tokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/connectors:
    get:
      tags: [Admin]
      summary: List all connectors
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Connectors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Connector'
    post:
      tags: [Admin]
      summary: Create new connector
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectorCreate'
      responses:
        '201':
          description: Connector created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connector'
  /admin/ingest/run:
    post:
      tags: [Admin]
      summary: Trigger manual ingestion run
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connector_name]
              properties:
                connector_name:
                  type: string
                  example: phoenix_legistar
      responses:
        '202':
          description: Ingestion started
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: running
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Address:
      type: object
      required: [line1, city, state, postal_code, country]
      properties:
        line1:
          type: string
          example: "200 W Washington St"
        line2:
          type: string
          example: ""
        city:
          type: string
          example: "Phoenix"
        state:
          type: string
          example: "AZ"
        postal_code:
          type: string
          example: "85003"
        country:
          type: string
          example: "US"
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
    Tokens:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    LocationResolution:
      type: object
      properties:
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
        divisions_resolved:
          type: boolean
...
    Connector:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        source:
          type: string
        enabled:
          type: boolean
        config:
          type: object
        updated_at:
          type: string
          format: date-time
    ConnectorCreate:
      type: object
      required: [name, source, config]
      properties:
        name:
          type: string
        source:
          type: string
          enum: [congress, govinfo, openstates, legiscan, legistar, custom]
        enabled:
          type: boolean
          default: true
        config:
          type: object
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'